requirements.txt
boto3>=1.34.0
botocore>=1.34.0
mcp>=1.0.0
python-dotenv>=1.0.0

# Development dependencies (optional)
# Install with: pip install -r requirements.txt -r requirements-dev.txt
# Or use: pip install -e ".[dev]" if you have setup.py
